% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{calculate_DE}
\alias{calculate_DE}
\title{Calculate sample-wise differential expression}
\usage{
calculate_DE(geneExpression)
}
\arguments{
\item{normalGeneExpression}{data.frame containing the gene expression in wide
format. A column \code{gene} indicates the gene, and the rest of the column
are the CPM for each sample.}

\item{tumorGeneExpression}{Equivalent to normalGeneExpression with tumor
data.}
}
\value{
A data.frame with the difference in expression \code{DE} between
the conditions per gene and sample.
}
\description{
Computes the gene expression difference between normal and tumor
conditions of the same sample. If no normal sample is available, computes the
difference with the median expression. Prior to that, performs a
normalization of the data.
}
