% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{calculate_dPSI}
\alias{calculate_dPSI}
\title{Calculate deltaPSI}
\usage{
calculate_dPSI(expression, tx2gene)
}
\arguments{
\item{expression}{data.frame with the expression information in long format,
with column names 'gene','sample','tpmCtrl' and 'tpmCase'.}

\item{tx2gene}{data.frame with two columns, named "transcript" and "gene",
which contains the transcript-gene mapping.}
}
\value{
A data.frame with the PSIs per sample and condition, and the
corresponding deltaPSI between conditions.
}
\description{
Calculates the deltaPSI between two conditions. When a subject
has matching samples for the two conditions, that value is taken. Else, the
deltaPSI is compared to the median of the controls.
}
