% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_ecdf}
\alias{compute_ecdf}
\title{Get empirical distribution}
\usage{
compute_ecdf(x, minValue, minSamples)
}
\arguments{
\item{x}{numerical vector}

\item{minValue}{minimum value to consider that measure}

\item{minSamples}{Minimum number of valid samples.}
}
\value{
\itemize{
\item {function returning NA when there are less than 10 valid values (ie non
NA and higher than threshold)}
\item {function returning 1 to any value higher than 0 when all the
differences are 0.}
\item {return ecdf otherwise, counting that the minimum p-value will be 1/n+1
}}
}
\description{
Get the empirical distribution of the differences between a
vector
}
