% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{find_switches}
\alias{find_switches}
\title{Find isoform switches}
\usage{
find_switches(txInfo)
}
\arguments{
\item{txInfo}{data.frame with transcript-level annotation per subject. Must
contain all of the following columns: \code{pDE} indicating if the gene is
differentially expressed; \code{dPSI} indicating the deltaPSI and
\code{pdPSI} scoring how extreme it is; \code{tpmNormal} and \code{tpmTumor}
indicating the expression in the two conditions.}
}
\value{
A data.frame with the measured isoform switches.
}
\description{
Find the isoform switches that match all of the following
conditions: 1) genes are not differentially expressed; 2) abs(deltaPSI) is > 0.05
and more extreme than 1% of the observed variance between controls; 3)
involved transcripts are sufficiently expressed (TPM > 0.1).
}
