% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGeneExpression}
\alias{getGeneExpression}
\title{Read expression}
\usage{
getGeneExpression(txExpression, tx2geneFile)
}
\arguments{
\item{txExpression}{Data frame with expression in TPM in long format.}

\item{tx2geneFile}{File containing genes-transcript correspondence, one
per line.}
}
\value{
Data frame with gene expression in TPM in long format.
}
\description{
Read expression
}
