% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize gene expression counts}
\usage{
normalize_counts(normalGeneExpression, tumorGeneExpression)
}
\arguments{
\item{normalGeneExpression}{data.frame containing the gene expression in wide
format. A column \code{gene} indicates the gene, and the rest of the column
are the CPM for each sample.}

\item{tumorGeneExpression}{Equivalent to normalGeneExpression with tumor
data.}
}
\value{
A data.frame with the normal and tumor gene expression (both if
samplenames match, else NA) per patient in log2 normalized counts.
}
\description{
Normalize gene expression counts.
}
