% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_tx_expression}
\alias{read_tx_expression}
\title{Read expression}
\usage{
read_tx_expression(ctrlExpressionFile, caseExpressionFile)
}
\arguments{
\item{ctrlExpressionFile}{File containing the transcript expression for
the control samples, measured in log2(TPMs + 0.001).}

\item{caseExpressionFile}{File containing the transcript expression for
the case samples, measured in log2(TPMs + 0.001).}
}
\value{
Data frame with transcript expression in TPM in long format.
}
\description{
Read expression
}
