% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{tpm2long}
\alias{tpm2long}
\title{Convert TPM to long format}
\usage{
tpm2long(wide, colname)
}
\arguments{
\item{wide}{Wide format data.frame with TPM values. First column contains the
transcript id, and succesive columns are the different samples.}

\item{colname}{Unquoted name of the desired expression column name.}
}
\value{
A data.frame with three columns: transcript, sample and expression.
}
\description{
Converts TPM, wide-format matrix to a long format tibble.
}
