% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uninformative_output}
\alias{uninformative_output}
\title{Pseudo-ecdf when input is insufficient}
\usage{
uninformative_output(x)
}
\arguments{
\item{x}{numerical vector.}
}
\value{
Numerical vector with NAs.
}
\description{
ecdf function where anything returns NA.
}
