% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uninformativeInput}
\alias{uninformativeInput}
\title{Pseudo-ecdf when all values are 0}
\usage{
uninformativeInput(x)
}
\arguments{
\item{x}{numerical vector}
}
\value{
Numerical vector with 0 when the input is 0, and 1 when it is > 0.
}
\description{
ecdf function where anything higher than 0 gets a p-value of 1.
}
