% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{compute_isoform_switches}
\alias{compute_isoform_switches}
\title{Compute isoform switches}
\usage{
compute_isoform_switches(ctrlExpressionFile, caseExpressionFile, tx2geneFile,
  outfile, minTpm = 0.1, minSamples = 10, ...)
}
\arguments{
\item{ctrlExpressionFile}{Tab-separated table with normalized
transcript-level expression for the control samples, in log2(TPM + 0.001).}

\item{caseExpressionFile}{Tab-separated table with normalized
transcript-level expression for the case samples, in log2(TPM + 0.001).}

\item{tx2geneFile}{Tab-separated table with the assignation of transcripts
to genes (one per line).}

\item{outfile}{Output file.}

\item{...}{arguments passed to the \code{\link{find_switches}} function.}
}
\description{
Top-level function to find isoform switches from transcript- and
gene-level expression files.
}
