% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{find_switches}
\alias{find_switches}
\title{Find isoform switches}
\usage{
find_switches(tx_pdPSI, gene_pDE, pDE = 0.01, pdPSI = 0.01,
  deltaPSI = 0.05, minExpression = 0.1)
}
\arguments{
\item{tx_pdPSI}{data.frame with transcript-level annotation per sample, as
outputted by \code{score_delta}. Must contain all of the following columns:
\code{dPSI} indicating the deltaPSI and \code{p} scoring how extreme it is;
\code{tpmCtrl} and \code{tpmCase} indicating the expression in the two
conditions; and \code{transcript}, \code{gene} and \code{sample}.}

\item{gene_pDE}{data.frame with gene-level annotation per sample, as
outputted by \code{score_delta}. Must contain all of the following columns:
\code{p} indicating if the gene is differentially expressed; and \code{gene}
and \code{sample}.}

\item{pDE}{Minimum p-value to consider a gene differentially expressed.}

\item{pdPSI}{Minimum p-value to consider a scplicing change significant.}

\item{deltaPSI}{Minimum PSI change to consider a scplicing change significant.}

\item{minExpression}{Minimum expression of a transcript to consider it
expressed.}
}
\value{
A data.frame with the measured isoform switches.
}
\description{
Find the isoform switches that match all of the following
conditions: 1) genes are not differentially expressed; 2) abs(deltaPSI) is
> 0.05 and more extreme than 1% of the observed variance between controls; 3)
involved transcripts are sufficiently expressed (TPM > 0.1).
}
