% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{score_delta}
\alias{score_delta}
\title{Measure p of delta}
\usage{
score_delta(df, probe, ctrl, delta, minTpm = 0, minSamples = 10)
}
\arguments{
\item{df}{data.frame containing the differences.}

\item{probe}{Unquoted name of the biomarker column (e.g. transcript).}

\item{ctrl}{Unquoted name of the expression in the control condition column
(e.g. tpmCtrl).}

\item{delta}{Unquoted name of the actual difference between conditions column
(e.g. dPSI).}

\item{minTpm}{Minimum TPM to consider a transcript/gene expressed.}

\item{minSamples}{Minimum number of samples to consider the differences
informative. If the number of samples with a valid value (i.e. >= minTpm)
is < minSamples, NAs will be returned.}
}
\value{
A copy of df with an extra column \code{p} indicating the probability
of the difference.
}
\description{
Assigns a probability of an observed difference between
conditions as function of the differences observed between the controls.
}
